Main: enum moved to main
