+++ src/main/java/pl/isa/thebuggers/fitly/Main.java
@@ -70,4 +70,23 @@ public class Main {
     private static void handleDietOption() {
         // TODO: implement
     }
+    private enum MenuOption {
+
+        DUMP(0), // do wygody
+        USER(1),
+        BMI(2),
+        TRAINING(3),
+        DIET(4),
+        EXIT(5);
+
+        private final int value;
+
+        MenuOption(int value) {
+            this.value = value;
+        }
+
+        public int getValue() {
+            return value;
+        }
+    }
 }
